创建直播功能完善，先走sdk，成功后走Api

+++ app/src/main/java/com/sunny/livechat/login/bean/UserBean.kt
@@ -13,7 +13,7 @@ class UserBean {
     var userId: Int? = null
     var username: String? = null
     var roleIdList: ArrayList<Int>? = null
-    var createTime: Int? = null
+    var createTime: Long? = null
     var deptId: Int? = null
     var deptName: String? = null
     var levelCode: String? = null
